IFEDMethodBase, part 7.
