+++ source/interaction/ifed_method.cc
@@ -1148,32 +1148,6 @@ namespace fdl
 
   template <int dim, int spacedim>
   void
-  IFEDMethod<dim, spacedim>::putToDatabase(tbox::Pointer<tbox::Database> db)
-  {
-    auto do_put = [&](auto &collection, const std::string &prefix)
-    {
-      for (unsigned int i = 0; i < collection.size(); ++i)
-        {
-          std::ostringstream              out_str;
-          boost::archive::binary_oarchive oarchive(out_str);
-          collection[i].save(oarchive, 0);
-          // TODO - with C++20 we can use view() instead of str() and skip
-          // this copy
-          const std::string out = out_str.str();
-          save_binary(prefix + std::to_string(i),
-                      out.c_str(),
-                      out.c_str() + out.size(),
-                      db);
-        }
-    };
-    do_put(this->parts, "part_");
-    do_put(this->surface_parts, "surface_part_");
-  }
-
-
-
-  template <int dim, int spacedim>
-  void
   IFEDMethod<dim, spacedim>::registerEulerianVariables()
   {
     // we need ghosts for CONSERVATIVE_LINEAR_REFINE
